# shared/hapi/model/hapi.gen.go
// Package model provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.15.0 DO NOT EDIT.
package model

import (
	"time"
)

// Defines values for AppConfigListSource.
const (
	AppConfigListSourceEnvironment AppConfigListSource = "environment"
	AppConfigListSourceStack       AppConfigListSource = "stack"
)

// Defines values for AppConfigReadSource.
const (
	AppConfigReadSourceEnvironment AppConfigReadSource = "environment"
	AppConfigReadSourceStack       AppConfigReadSource = "stack"
)

// AppConfigList defines model for AppConfigList.
type AppConfigList struct {
	AppName     string              `json:"app_name"`
	CreatedAt   time.Time           `json:"created_at"`
	DeletedAt   *time.Time          `json:"deleted_at,omitempty"`
	Environment string              `json:"environment"`
	Id          int64               `json:"id"`
	Key         string              `json:"key"`
	Source      AppConfigListSource `json:"source"`
	Stack       *string             `json:"stack,omitempty"`
	UpdatedAt   time.Time           `json:"updated_at"`
	Value       string              `json:"value"`
}

// AppConfigListSource defines model for AppConfigList.Source.
type AppConfigListSource string

// AppConfigRead defines model for AppConfigRead.
type AppConfigRead struct {
	AppName     string              `json:"app_name"`
	CreatedAt   time.Time           `json:"created_at"`
	DeletedAt   *time.Time          `json:"deleted_at,omitempty"`
	Environment string              `json:"environment"`
	Id          int64               `json:"id"`
	Key         string              `json:"key"`
	Source      AppConfigReadSource `json:"source"`
	Stack       *string             `json:"stack,omitempty"`
	UpdatedAt   time.Time           `json:"updated_at"`
	Value       string              `json:"value"`
}

// AppConfigReadSource defines model for AppConfigRead.Source.
type AppConfigReadSource string

// N400 defines model for 400.
type N400 struct {
	Code   int          `json:"code"`
	Errors *interface{} `json:"errors,omitempty"`
	Status string       `json:"status"`
}

// N404 defines model for 404.
type N404 struct {
	Code   int          `json:"code"`
	Errors *interface{} `json:"errors,omitempty"`
	Status string       `json:"status"`
}

// N409 defines model for 409.
type N409 struct {
	Code   int          `json:"code"`
	Errors *interface{} `json:"errors,omitempty"`
	Status string       `json:"status"`
}

// N500 defines model for 500.
type N500 struct {
	Code   int          `json:"code"`
	Errors *interface{} `json:"errors,omitempty"`
	Status string       `json:"status"`
}

// ListAppConfigParams defines parameters for ListAppConfig.
type ListAppConfigParams struct {
	// Page what page to render
	Page *int `form:"page,omitempty" json:"page,omitempty"`

	// ItemsPerPage item count to render per page
	ItemsPerPage *int `form:"itemsPerPage,omitempty" json:"itemsPerPage,omitempty"`
}